requirements.txt
attrs==23.2.0
blinker==1.7.0
certifi==2023.11.17
chardet==4.0.0
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
dataclasses==0.6
defusedxml==0.7.1
deprecation==2.1.0
Flask==3.0.0
Flask-Mail==0.9.1
fonttools==4.47.0
fpdf2==2.7.7
gotrue==0.2.0
gunicorn==21.2.0
h11==0.14.0
httpcore==0.12.3
httpx==0.16.1
idna==2.10
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
numpy==1.26.3
packaging==23.2
pandas==2.1.4
pdfkit==1.0.0
pillow==10.2.0
plotly==5.18.0
pluggy==1.3.0
postgrest-py==0.4.0
py==1.11.0
pyparsing==3.1.1
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
realtime-py==0.1.3
requests==2.25.1
rfc3986==1.5.0
seaborn==0.13.1
six==1.16.0
sniffio==1.3.0
supabase-py==0.0.2
tenacity==8.2.3
toml==0.10.2
tzdata==2023.4
urllib3==1.26.18
websockets==9.1
Werkzeug==3.0.1
